requirements.txt
Flask
scikit-learn
scipy
numpy
pandas
sklearn
gunicorn
tensorflow-cpu
Pillow